requirements.txt
streamlit
requests
langchain-community
faiss-cpu
sentence-transformers
python-dotenv